backend/requirements.txt
# Flask Framework and Extensions
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-CORS==4.0.0
Flask-Session==0.5.0

# Database Support
SQLAlchemy==2.0.20
psycopg2-binary==2.9.7
alembic==1.12.0

# LDAP/Active Directory Authentication
python-ldap==3.4.3
ldap3==2.9.1

# Security and Validation
Werkzeug==2.3.7
WTForms==3.0.1
email-validator==2.0.0
cryptography==41.0.4

# File Processing and Export
Pillow==10.0.0
python-magic==0.4.27
reportlab==4.0.4
markdown==3.4.4
python-markdown-math==0.8
pymdown-extensions==10.2.1

# Full-Text Search
whoosh==2.7.4
elasticsearch==8.9.0

# Utilities and Helpers
python-dotenv==1.0.0
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
urllib3==2.0.4

# Development and Testing
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==19.6.1

# Production Server
gunicorn==21.2.0
gevent==23.7.0

# API Documentation
flask-restx==1.1.0
flask-apispec==0.11.4
apispec==6.3.0
marshmallow==3.20.1

# Monitoring and Logging
structlog==23.1.0
colorama==0.4.6

# File Format Support
PyPDF2==3.0.1
openpyxl==3.1.2
python-docx==0.8.11

# Backup and Export
schedule==1.2.0
zipfile36==0.1.3

# Environment and Configuration
environs==10.0.0
dynaconf==3.2.2